refactor: split runner.rs into submodules (#102)
Extract from runner/mod.rs into:
- project_discovery.rs: ValidProject, discover_and_validate, build_plan_files
- plan.rs: build_plan_request
- artifact_pipeline.rs: ArtifactEntry, run_artifact_pipeline
- manifest_builder.rs: build_diff_and_manifest, build_manifest_checks
- submission.rs: create_board_run, submit_bundle

mod.rs retains run() and process_project() as thin orchestrators.
