+++ crates/action-runner/src/runner/mod.rs
@@ -1,20 +1,13 @@
-use std::fs;
-use std::path::{Path, PathBuf};
+mod artifact_pipeline;
+mod manifest_builder;
+mod plan;
+mod project_discovery;
+mod submission;
 
-use boardflow_domain::models::artifact::ArtifactType;
-use boardflow_domain::public_ids::BoardProjectId;
-use serde::Serialize;
-use tracing::{error, info, warn};
-
-use boardflow_api_types::board_run::{CreateBoardRunRequest, ImportArtifactBundleRequest};
-use boardflow_api_types::plan::{
-    PlanActionInput, PlanDecision, PlanGitInput, PlanMode, PlanProjectFile, PlanProjectInput,
-    PlanRepositoryInput, PlanRequest,
-};
-use boardflow_kicad::cli::{KicadCli, PcbSide};
-use boardflow_kicad::config::{self, BoardflowConfig};
-use boardflow_kicad::detect;
+use boardflow_api_types::plan::PlanDecision;
+use boardflow_kicad::cli::KicadCli;
 use boardflow_kicad::hash;
+use tracing::{error, info};
 
 use crate::api::ApiClient;
 use crate::bundle;
@@ -22,91 +15,7 @@ use crate::error::ActionError;
 use crate::inputs::{self, ActionInputs, GitHubContext};
 use crate::summary::{self, ProjectResult};
 
-struct ValidProject {
-    project_dir: PathBuf,
-    pro_file: PathBuf,
-    pcb_file: PathBuf,
-    sch_file: PathBuf,
-    excludes: Vec<String>,
-    #[allow(dead_code)]
-    config: BoardflowConfig,
-    rel_dir: String,
-    rel_pro_path: String,
-}
-
-#[derive(Serialize)]
-struct ArtifactEntry {
-    #[serde(rename = "type")]
-    artifact_type: ArtifactType,
-    status: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    path: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    source_path: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    content_type: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    sha256: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    size_bytes: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    error_message: Option<String>,
-}
-
-impl ArtifactEntry {
-    fn available(
-        artifact_type: ArtifactType,
-        path: &str,
-        content_type: &str,
-        file_path: &Path,
-    ) -> Self {
-        let sha256 = hash::compute_file_sha256(file_path).ok();
-        let size_bytes = fs::metadata(file_path).ok().map(|m| m.len());
-        Self {
-            artifact_type,
-            status: "available".to_string(),
-            path: Some(path.to_string()),
-            source_path: None,
-            content_type: Some(content_type.to_string()),
-            sha256: sha256.map(|h| format!("sha256:{h}")),
-            size_bytes,
-            error_message: None,
-        }
-    }
-
-    fn failed(artifact_type: ArtifactType, message: &str) -> Self {
-        Self {
-            artifact_type,
-            status: "failed".to_string(),
-            path: None,
-            source_path: None,
-            content_type: None,
-            sha256: None,
-            size_bytes: None,
-            error_message: Some(message.to_string()),
-        }
-    }
-
-    fn source(
-        artifact_type: ArtifactType,
-        staging_path: &str,
-        source_path: &str,
-        file_path: &Path,
-    ) -> Self {
-        let sha256 = hash::compute_file_sha256(file_path).ok();
-        let size_bytes = fs::metadata(file_path).ok().map(|m| m.len());
-        Self {
-            artifact_type,
-            status: "available".to_string(),
-            path: Some(staging_path.to_string()),
-            source_path: Some(source_path.to_string()),
-            content_type: None,
-            sha256: sha256.map(|h| format!("sha256:{h}")),
-            size_bytes,
-            error_message: None,
-        }
-    }
-}
+use project_discovery::ValidProject;
 
 pub async fn run() -> i32 {
     // 1. Parse inputs
@@ -117,7 +26,6 @@ pub async fn run() -> i32 {
             return 1;
         }
     };
-
     let gh = match inputs::parse_github_context() {
         Ok(g) => g,
         Err(e) => {
@@ -142,151 +50,9 @@ pub async fn run() -> i32 {
         return 0;
     }
 
-    // 3. Detect projects
-    let yml_paths = match detect::find_boardflow_ymls(&gh.workspace) {
-        Ok(p) => p,
-        Err(_) => {
-            summary::error("No .boardflow.yml found in workspace");
-            return 1;
-        }
-    };
-
-    // 4. Validate projects
-    let mut valid_projects: Vec<ValidProject> = Vec::new();
-    let mut detection_errors = 0u32;
-
-    for yml_path in &yml_paths {
-        let project_dir = match yml_path.parent() {
-            Some(d) => d.to_path_buf(),
-            None => {
-                detection_errors += 1;
-                continue;
-            }
-        };
-
-        let cfg = match config::parse_boardflow_yml(yml_path) {
-            Ok(c) => c,
-            Err(e) => {
-                summary::warning(&format!("Failed to parse {}: {e}", yml_path.display()));
-                detection_errors += 1;
-                continue;
-            }
-        };
-
-        if let Err(e) = config::validate_schema_v1(&cfg) {
-            summary::warning(&format!("Invalid schema in {}: {e}", yml_path.display()));
-            detection_errors += 1;
-            continue;
-        }
-
-        // Merge excludes (action.yml specifies newline-separated patterns)
-        let input_excludes: Vec<String> = action_inputs
-            .exclude_paths
-            .lines()
-            .map(|s| s.trim().to_string())
-            .filter(|s| !s.is_empty())
-            .collect();
-        let excludes =
-            config::merge_excludes(hash::BUILTIN_EXCLUDES, &input_excludes, &cfg.exclude_paths);
-
-        let pro_file = match detect::resolve_kicad_pro(&project_dir) {
-            Ok(p) => p,
-            Err(e) => {
-                summary::warning(&format!(
-                    "No unique .kicad_pro in {}: {e}",
-                    project_dir.display()
-                ));
-                detection_errors += 1;
-                continue;
-            }
-        };
-
-        let pro_stem = pro_file
-            .file_stem()
-            .and_then(|s| s.to_str())
-            .unwrap_or_default()
-            .to_string();
-
-        let pcb_file = match detect::resolve_pcb_file(&project_dir, &pro_stem) {
-            Ok(p) => p,
-            Err(e) => {
-                summary::warning(&format!(
-                    "No .kicad_pcb for {pro_stem} in {}: {e}",
-                    project_dir.display()
-                ));
-                detection_errors += 1;
-                continue;
-            }
-        };
-
-        let sch_file = match detect::resolve_root_schematic(&project_dir, &pro_stem) {
-            Ok(p) => p,
-            Err(e) => {
-                summary::warning(&format!(
-                    "No .kicad_sch for {pro_stem} in {}: {e}",
-                    project_dir.display()
-                ));
-                detection_errors += 1;
-                continue;
-            }
-        };
-
-        // Validate required files are not excluded
-        let pro_rel = pro_file
-            .file_name()
-            .unwrap_or_default()
-            .to_str()
-            .unwrap_or_default();
-        let pcb_rel = pcb_file
-            .file_name()
-            .unwrap_or_default()
-            .to_str()
-            .unwrap_or_default();
-        let sch_rel = sch_file
-            .file_name()
-            .unwrap_or_default()
-            .to_str()
-            .unwrap_or_default();
-        if hash::is_excluded(pro_rel, &excludes)
-            || hash::is_excluded(pcb_rel, &excludes)
-            || hash::is_excluded(sch_rel, &excludes)
-        {
-            summary::warning(&format!(
-                "Required files excluded in {}",
-                project_dir.display()
-            ));
-            detection_errors += 1;
-            continue;
-        }
-
-        // Compute relative paths
-        let rel_dir = project_dir
-            .strip_prefix(&gh.workspace)
-            .map(|p| p.to_str().unwrap_or("."))
-            .unwrap_or(".")
-            .to_string();
-        let rel_dir = if rel_dir.is_empty() {
-            ".".to_string()
-        } else {
-            rel_dir
-        };
-
-        let rel_pro_path = pro_file
-            .strip_prefix(&gh.workspace)
-            .map(|p| p.to_str().unwrap_or_default().to_string())
-            .unwrap_or_default();
-
-        valid_projects.push(ValidProject {
-            project_dir,
-            pro_file,
-            pcb_file,
-            sch_file,
-            excludes,
-            config: cfg,
-            rel_dir,
-            rel_pro_path,
-        });
-    }
+    // 3-4. Discover and validate projects
+    let (valid_projects, detection_errors) =
+        project_discovery::discover_and_validate(&gh.workspace, &action_inputs);
 
     if valid_projects.is_empty() {
         summary::error("No valid projects found");
@@ -294,65 +60,7 @@ pub async fn run() -> i32 {
     }
 
     // 5. Compute hashes and build plan payload
-    let mut plan_projects = Vec::new();
-    for vp in &valid_projects {
-        let tree_hash = match hash::compute_tree_hash(&vp.project_dir, &vp.excludes) {
-            Ok(h) => h,
-            Err(e) => {
-                summary::warning(&format!(
-                    "Failed to compute tree hash for {}: {e}",
-                    vp.rel_dir
-                ));
-                continue;
-            }
-        };
-
-        let files = build_plan_files(&vp.project_dir, &vp.excludes);
-        let yml_rel = if vp.rel_dir == "." {
-            ".boardflow.yml".to_string()
-        } else {
-            format!("{}/.boardflow.yml", vp.rel_dir)
-        };
-
-        plan_projects.push(PlanProjectInput {
-            project_path: vp.rel_pro_path.clone(),
-            config_path: yml_rel,
-            project_dir: vp.rel_dir.clone(),
-            tree_hash: format!("sha256:{tree_hash}"),
-            files,
-        });
-    }
-
-    let mode = match action_inputs.mode.as_str() {
-        "all" => PlanMode::All,
-        _ => PlanMode::Auto,
-    };
-
-    let github_repository_id = std::env::var("GITHUB_REPOSITORY_ID").unwrap_or_default();
-
-    let plan_request = PlanRequest {
-        repository: PlanRepositoryInput {
-            github_repository_id,
-            owner: gh.owner.clone(),
-            name: gh.repo_name.clone(),
-        },
-        git: PlanGitInput {
-            ref_: gh.git_ref.clone(),
-            branch: gh.ref_name.clone(),
-            commit_sha: gh.sha.clone(),
-            event_name: gh.event_name.clone(),
-        },
-        action: PlanActionInput {
-            workflow: "BoardFlow".to_string(),
-            run_id: gh.run_id.clone(),
-            run_attempt: gh.run_attempt.clone(),
-        },
-        mode,
-        projects: plan_projects,
-    };
-
-    let plan_payload =
-        serde_json::to_value(&plan_request).expect("failed to serialize plan request");
+    let plan_payload = plan::build_plan_request(&valid_projects, &action_inputs, &gh);
 
     // 6. Call plan API
     let api = ApiClient::new(&action_inputs.api_url, &action_inputs.token);
@@ -448,7 +156,7 @@ async fn process_project(
     vp: &ValidProject,
     gh: &GitHubContext,
     inputs: &ActionInputs,
-    board_project_id: BoardProjectId,
+    board_project_id: boardflow_domain::public_ids::BoardProjectId,
 ) -> std::result::Result<bool, ActionError> {
     let pro_stem = vp
         .pro_file
@@ -464,22 +172,7 @@ async fn process_project(
     let output_path = output_dir.path();
 
     // Create board run
-    let tree_hash = hash::compute_tree_hash(&vp.project_dir, &vp.excludes)
-        .map_err(|e| ActionError::Bundle(format!("tree hash: {e}")))?;
-
-    let create_payload = serde_json::to_value(&CreateBoardRunRequest {
-        board_project_id,
-        project_path: vp.rel_pro_path.clone(),
-        tree_hash: format!("sha256:{tree_hash}"),
-        commit_sha: gh.sha.clone(),
-        branch: gh.ref_name.clone(),
-        ref_: gh.git_ref.clone(),
-        github_run_id: gh.run_id.clone(),
-        github_run_attempt: gh.run_attempt.clone(),
-    })
-    .expect("failed to serialize create_board_run request");
-
-    let create_resp = api.create_board_run(&create_payload).await?;
+    let create_resp = submission::create_board_run(api, vp, gh, board_project_id).await?;
     let board_run_id = create_resp.board_run_id;
 
     // If artifact_bundle is None, the run already exists in a terminal or importing state.
@@ -497,565 +190,19 @@ async fn process_project(
     let upload_url = &artifact_bundle.upload_url;
     let staging_object_key = &artifact_bundle.object_key;
 
-    let mut artifacts: Vec<serde_json::Value> = Vec::new();
-    let mut checks_failed = false;
+    // Run artifact pipeline (ERC, DRC, PDFs, SVGs, Gerber, Drill, BOM, Position, 3D, iBOM, source files)
+    let (artifacts, checks_failed) =
+        artifact_pipeline::run_artifact_pipeline(kicad, vp, inputs, output_path).await?;
 
-    // Run ERC
-    let erc_json = output_path.join("erc.json");
-    match kicad.run_erc(&vp.sch_file, &erc_json).await {
-        Ok(cmd_out) => {
-            if cmd_out.exit_code == 0 || cmd_out.exit_code == 5 {
-                artifacts.push(
-                    serde_json::to_value(ArtifactEntry::available(
-                        ArtifactType::ErcReport,
-                        "checks/erc.json",
-                        "application/json",
-                        &erc_json,
-                    ))
-                    .unwrap(),
-                );
-                if cmd_out.exit_code == 5 && inputs.fail_on_erc {
-                    checks_failed = true;
-                }
-            } else {
-                artifacts.push(
-                    serde_json::to_value(ArtifactEntry::failed(
-                        ArtifactType::ErcReport,
-                        "ERC execution failed",
-                    ))
-                    .unwrap(),
-                );
-            }
-        }
-        Err(e) => {
-            warn!("ERC failed: {e}");
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::failed(
-                    ArtifactType::ErcReport,
-                    &format!("ERC execution failed: {e}"),
-                ))
-                .unwrap(),
-            );
-        }
-    }
-
-    // Run DRC
-    let drc_json = output_path.join("drc.json");
-    match kicad.run_drc(&vp.pcb_file, &drc_json).await {
-        Ok(cmd_out) => {
-            if cmd_out.exit_code == 0 || cmd_out.exit_code == 5 {
-                artifacts.push(
-                    serde_json::to_value(ArtifactEntry::available(
-                        ArtifactType::DrcReport,
-                        "checks/drc.json",
-                        "application/json",
-                        &drc_json,
-                    ))
-                    .unwrap(),
-                );
-                if cmd_out.exit_code == 5 && inputs.fail_on_drc {
-                    checks_failed = true;
-                }
-            } else {
-                artifacts.push(
-                    serde_json::to_value(ArtifactEntry::failed(
-                        ArtifactType::DrcReport,
-                        "DRC execution failed",
-                    ))
-                    .unwrap(),
-                );
-            }
-        }
-        Err(e) => {
-            warn!("DRC failed: {e}");
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::failed(
-                    ArtifactType::DrcReport,
-                    &format!("DRC execution failed: {e}"),
-                ))
-                .unwrap(),
-            );
-        }
-    }
-
-    // Export PCB PDF
-    let pdf_dir = output_path.join("pdf");
-    fs::create_dir_all(&pdf_dir)?;
-    match kicad
-        .export_pcb_pdf(&vp.pcb_file, &pdf_dir.join("pcb.pdf"))
-        .await
-    {
-        Ok(_) => {
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::available(
-                    ArtifactType::PcbPdf,
-                    "review/pcb.pdf",
-                    "application/pdf",
-                    &pdf_dir.join("pcb.pdf"),
-                ))
-                .unwrap(),
-            );
-        }
-        Err(e) => {
-            warn!("PCB PDF failed: {e}");
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::failed(
-                    ArtifactType::PcbPdf,
-                    "PCB PDF export failed",
-                ))
-                .unwrap(),
-            );
-        }
-    }
-
-    // Export Schematic PDF
-    match kicad
-        .export_sch_pdf(&vp.sch_file, &pdf_dir.join("schematic.pdf"))
-        .await
-    {
-        Ok(_) => {
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::available(
-                    ArtifactType::SchematicPdf,
-                    "review/schematic.pdf",
-                    "application/pdf",
-                    &pdf_dir.join("schematic.pdf"),
-                ))
-                .unwrap(),
-            );
-        }
-        Err(e) => {
-            warn!("Schematic PDF failed: {e}");
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::failed(
-                    ArtifactType::SchematicPdf,
-                    "Schematic PDF export failed",
-                ))
-                .unwrap(),
-            );
-        }
-    }
-
-    // Export SVG
-    let svg_dir = output_path.join("svg");
-    fs::create_dir_all(&svg_dir)?;
-    match kicad
-        .export_pcb_svg(&vp.pcb_file, &svg_dir.join("pcb_top.svg"), PcbSide::Top)
-        .await
-    {
-        Ok(_) => {
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::available(
-                    ArtifactType::PcbTopSvg,
-                    "review/pcb_top.svg",
-                    "image/svg+xml",
-                    &svg_dir.join("pcb_top.svg"),
-                ))
-                .unwrap(),
-            );
-        }
-        Err(e) => {
-            warn!("PCB top SVG failed: {e}");
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::failed(
-                    ArtifactType::PcbTopSvg,
-                    "PCB top SVG export failed",
-                ))
-                .unwrap(),
-            );
-        }
-    }
-
-    match kicad
-        .export_pcb_svg(
-            &vp.pcb_file,
-            &svg_dir.join("pcb_bottom.svg"),
-            PcbSide::Bottom,
-        )
-        .await
-    {
-        Ok(_) => {
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::available(
-                    ArtifactType::PcbBottomSvg,
-                    "review/pcb_bottom.svg",
-                    "image/svg+xml",
-                    &svg_dir.join("pcb_bottom.svg"),
-                ))
-                .unwrap(),
-            );
-        }
-        Err(e) => {
-            warn!("PCB bottom SVG failed: {e}");
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::failed(
-                    ArtifactType::PcbBottomSvg,
-                    "PCB bottom SVG export failed",
-                ))
-                .unwrap(),
-            );
-        }
-    }
-
-    // Export Gerber
-    let gerber_dir = output_path.join("gerber");
-    fs::create_dir_all(&gerber_dir)?;
-    let gerber_ok = kicad
-        .export_gerbers(&vp.pcb_file, &gerber_dir)
-        .await
-        .is_ok();
-
-    // Export Drill
-    let drill_dir = output_path.join("drill");
-    fs::create_dir_all(&drill_dir)?;
-    let drill_ok = kicad.export_drill(&vp.pcb_file, &drill_dir).await.is_ok();
-
-    // Create zip archives
-    let gerbers_zip = output_path.join("gerbers.zip");
-    if gerber_ok {
-        if bundle::create_bundle_zip(&gerber_dir, &gerbers_zip).is_ok() {
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::available(
-                    ArtifactType::GerberZip,
-                    "fabrication/gerbers.zip",
-                    "application/zip",
-                    &gerbers_zip,
-                ))
-                .unwrap(),
-            );
-        } else {
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::failed(
-                    ArtifactType::GerberZip,
-                    "Gerber zip creation failed",
-                ))
-                .unwrap(),
-            );
-        }
-    } else {
-        artifacts.push(
-            serde_json::to_value(ArtifactEntry::failed(
-                ArtifactType::GerberZip,
-                "Gerber export failed",
-            ))
-            .unwrap(),
-        );
-    }
-
-    let drill_zip = output_path.join("drill.zip");
-    if drill_ok {
-        if bundle::create_bundle_zip(&drill_dir, &drill_zip).is_ok() {
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::available(
-                    ArtifactType::DrillZip,
-                    "fabrication/drill.zip",
-                    "application/zip",
-                    &drill_zip,
-                ))
-                .unwrap(),
-            );
-        } else {
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::failed(
-                    ArtifactType::DrillZip,
-                    "Drill zip creation failed",
-                ))
-                .unwrap(),
-            );
-        }
-    } else {
-        artifacts.push(
-            serde_json::to_value(ArtifactEntry::failed(
-                ArtifactType::DrillZip,
-                "Drill export failed",
-            ))
-            .unwrap(),
-        );
-    }
-
-    // Fabrication zip (combined)
-    let fab_zip = output_path.join("fabrication.zip");
-    if gerber_ok || drill_ok {
-        if bundle::create_fabrication_zip(&gerber_dir, &drill_dir, &fab_zip).is_ok() {
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::available(
-                    ArtifactType::FabricationZip,
-                    "fabrication/fabrication.zip",
-                    "application/zip",
-                    &fab_zip,
-                ))
-                .unwrap(),
-            );
-        } else {
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::failed(
-                    ArtifactType::FabricationZip,
-                    "Fabrication zip creation failed",
-                ))
-                .unwrap(),
-            );
-        }
-    } else {
-        artifacts.push(
-            serde_json::to_value(ArtifactEntry::failed(
-                ArtifactType::FabricationZip,
-                "Fabrication zip creation failed",
-            ))
-            .unwrap(),
-        );
-    }
-
-    // Export BOM
-    let bom_dir = output_path.join("bom");
-    fs::create_dir_all(&bom_dir)?;
-    match kicad
-        .export_bom(&vp.sch_file, &bom_dir.join("bom.csv"))
-        .await
-    {
-        Ok(_) => {
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::available(
-                    ArtifactType::BomCsv,
-                    "assembly/bom.csv",
-                    "text/csv",
-                    &bom_dir.join("bom.csv"),
-                ))
-                .unwrap(),
-            );
-        }
-        Err(e) => {
-            warn!("BOM export failed: {e}");
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::failed(
-                    ArtifactType::BomCsv,
-                    "BOM export failed",
-                ))
-                .unwrap(),
-            );
-        }
-    }
-
-    // Export Position
-    let pos_dir = output_path.join("position");
-    fs::create_dir_all(&pos_dir)?;
-    match kicad
-        .export_position(&vp.pcb_file, &pos_dir.join("position.csv"))
-        .await
-    {
-        Ok(_) => {
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::available(
-                    ArtifactType::PositionCsv,
-                    "assembly/position.csv",
-                    "text/csv",
-                    &pos_dir.join("position.csv"),
-                ))
-                .unwrap(),
-            );
-        }
-        Err(e) => {
-            warn!("Position export failed: {e}");
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::failed(
-                    ArtifactType::PositionCsv,
-                    "Position export failed",
-                ))
-                .unwrap(),
-            );
-        }
-    }
-
-    // 3D Renders
-    let render_dir = output_path.join("3d");
-    fs::create_dir_all(&render_dir)?;
-    match kicad
-        .render_3d(&vp.pcb_file, &render_dir.join("top.png"), PcbSide::Top)
-        .await
-    {
-        Ok(_) => {
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::available(
-                    ArtifactType::RenderTopPng,
-                    "review/render_top.png",
-                    "image/png",
-                    &render_dir.join("top.png"),
-                ))
-                .unwrap(),
-            );
-        }
-        Err(e) => {
-            warn!("3D top render failed: {e}");
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::failed(
-                    ArtifactType::RenderTopPng,
-                    "3D top render failed",
-                ))
-                .unwrap(),
-            );
-        }
-    }
-
-    match kicad
-        .render_3d(
-            &vp.pcb_file,
-            &render_dir.join("bottom.png"),
-            PcbSide::Bottom,
-        )
-        .await
-    {
-        Ok(_) => {
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::available(
-                    ArtifactType::RenderBottomPng,
-                    "review/render_bottom.png",
-                    "image/png",
-                    &render_dir.join("bottom.png"),
-                ))
-                .unwrap(),
-            );
-        }
-        Err(e) => {
-            warn!("3D bottom render failed: {e}");
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::failed(
-                    ArtifactType::RenderBottomPng,
-                    "3D bottom render failed",
-                ))
-                .unwrap(),
-            );
-        }
-    }
-
-    // iBOM
-    let ibom_dir = output_path.join("ibom");
-    fs::create_dir_all(&ibom_dir)?;
-    match boardflow_kicad::ibom::run_ibom(&vp.pcb_file, &ibom_dir).await {
-        Ok(html_path) => {
-            // Copy to expected staging name
-            let dest = ibom_dir.join("ibom.html");
-            if html_path != dest {
-                let _ = fs::copy(&html_path, &dest);
-            }
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::available(
-                    ArtifactType::Ibom,
-                    "assembly/ibom.html",
-                    "text/html",
-                    &dest,
-                ))
-                .unwrap(),
-            );
-        }
-        Err(e) => {
-            warn!("iBOM failed: {e}");
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::failed(
-                    ArtifactType::Ibom,
-                    "iBOM generation failed",
-                ))
-                .unwrap(),
-            );
-        }
-    }
-
-    // KiCad source files as artifacts
-    if let Ok(entries) = fs::read_dir(&vp.project_dir) {
-        let mut src_entries: Vec<_> = entries
-            .flatten()
-            .filter(|e| {
-                e.path().is_file()
-                    && matches!(
-                        e.path().extension().and_then(|e| e.to_str()),
-                        Some("kicad_pro" | "kicad_sch" | "kicad_pcb" | "kicad_wks")
-                    )
-            })
-            .collect();
-        src_entries.sort_by_key(|e| e.path());
-
-        for entry in src_entries {
-            let path = entry.path();
-            let file_name = path
-                .file_name()
-                .and_then(|n| n.to_str())
-                .unwrap_or_default();
-            let src_rel = file_name.to_string();
-
-            if hash::is_excluded(&src_rel, &vp.excludes) {
-                continue;
-            }
-
-            let kicad_type = match path.extension().and_then(|e| e.to_str()) {
-                Some("kicad_pro") => ArtifactType::KicadPro,
-                Some("kicad_sch") => ArtifactType::KicadSch,
-                Some("kicad_pcb") => ArtifactType::KicadPcb,
-                Some("kicad_wks") => ArtifactType::KicadWks,
-                _ => continue,
-            };
-
-            let (staging_path, source_path) = if vp.rel_dir == "." {
-                (format!("kicad/{src_rel}"), src_rel.clone())
-            } else {
-                (
-                    format!("kicad/{}/{src_rel}", vp.rel_dir),
-                    format!("{}/{src_rel}", vp.rel_dir),
-                )
-            };
-
-            artifacts.push(
-                serde_json::to_value(ArtifactEntry::source(
-                    kicad_type,
-                    &staging_path,
-                    &source_path,
-                    &path,
-                ))
-                .unwrap(),
-            );
-        }
-    }
-
-    // Generate diff metadata
-    let diff_dir = output_path.join("diff");
-    fs::create_dir_all(&diff_dir)?;
-    let _ = bundle::generate_file_hashes_json(
-        &vp.project_dir,
-        &vp.excludes,
-        &diff_dir.join("file_hashes.json"),
-    );
-    let _ = bundle::generate_bom_summary_json(
-        &bom_dir.join("bom.csv"),
-        &diff_dir.join("bom_summary.json"),
-    );
-    let _ = bundle::generate_checks_summary_json(
-        &erc_json,
-        &drc_json,
-        &diff_dir.join("checks_summary.json"),
-    );
-    let _ = bundle::generate_artifacts_summary_json(
-        &artifacts,
-        &diff_dir.join("artifacts_summary.json"),
-    );
-    let _ = bundle::generate_previews_json(output_path, &diff_dir.join("previews.json"));
-
-    // Build manifest checks from ERC/DRC reports
-    let manifest_checks = build_manifest_checks(&erc_json, &drc_json);
-
-    // Build manifest files from project directory
-    let manifest_files: Vec<serde_json::Value> = build_plan_files(&vp.project_dir, &vp.excludes)
-        .into_iter()
-        .map(|f| serde_json::json!({"path": f.path, "sha256": f.sha256}))
-        .collect();
-
-    // Create manifest
-    let manifest_path = output_path.join("manifest.json");
-    bundle::create_manifest(
-        &vp.rel_pro_path,
-        &format!("sha256:{tree_hash}"),
+    // Build diff metadata + manifest
+    let tree_hash = hash::compute_tree_hash(&vp.project_dir, &vp.excludes)
+        .map_err(|e| ActionError::Bundle(format!("tree hash: {e}")))?;
+    let manifest_path = manifest_builder::build_diff_and_manifest(
+        vp,
+        output_path,
+        &tree_hash,
         &gh.sha,
-        &diff_dir.join("checks_summary.json"),
         &artifacts,
-        &manifest_checks,
-        &manifest_files,
-        &manifest_path,
     )?;
 
     // Build staging directory
@@ -1067,243 +214,17 @@ async fn process_project(
         &manifest_path,
     )?;
 
-    // Create bundle zip
-    let bundle_path = output_path.join("bundle.zip");
-    if let Err(e) = bundle::create_bundle_zip(&staging_dir, &bundle_path) {
-        let _ = api
-            .fail(board_run_id, "Bundle creation failed", &e.to_string())
-            .await;
-        return Err(ActionError::Bundle(format!("Bundle creation failed: {e}")));
-    }
-
-    let bundle_sha256 = bundle::compute_bundle_sha256(&bundle_path)?;
-
-    // Upload bundle
-    if let Err(e) = api.upload_bundle(upload_url, &bundle_path).await {
-        let _ = api
-            .fail(board_run_id, "Upload failed", &e.to_string())
-            .await;
-        return Err(e);
-    }
-
-    // Import
-    let bundle_size = fs::metadata(&bundle_path)?.len();
-    let import_payload = serde_json::to_value(&ImportArtifactBundleRequest {
-        staging_object_key: staging_object_key.clone(),
-        bundle_sha256: format!("sha256:{bundle_sha256}"),
-        bundle_size_bytes: bundle_size as i64,
-    })
-    .expect("failed to serialize import request");
-
-    if let Err(e) = api.import(board_run_id, &import_payload).await {
-        let _ = api
-            .fail(board_run_id, "Import failed", &e.to_string())
-            .await;
-        return Err(e);
-    }
+    // Submit bundle (zip, upload, import)
+    submission::submit_bundle(
+        api,
+        board_run_id,
+        &staging_dir,
+        output_path,
+        upload_url,
+        staging_object_key,
+    )
+    .await?;
 
     info!("Successfully processed project: {}", vp.rel_pro_path);
     Ok(checks_failed)
-}
-
-fn build_plan_files(project_dir: &Path, excludes: &[String]) -> Vec<PlanProjectFile> {
-    let files = match hash::list_project_files(project_dir, excludes) {
-        Ok(f) => f,
-        Err(_) => return Vec::new(),
-    };
-
-    let mut entries: Vec<(String, PathBuf)> = files
-        .into_iter()
-        .filter_map(|abs_path| {
-            let rel = abs_path.strip_prefix(project_dir).ok()?;
-            Some((rel.to_str()?.to_string(), abs_path))
-        })
-        .collect();
-    entries.sort_by(|a, b| a.0.cmp(&b.0));
-
-    entries
-        .into_iter()
-        .filter_map(|(rel_path, abs_path)| {
-            let file_hash = hash::compute_file_sha256(&abs_path).ok()?;
-            Some(PlanProjectFile {
-                path: rel_path,
-                sha256: format!("sha256:{file_hash}"),
-            })
-        })
-        .collect()
-}
-
-/// Build ManifestCheck entries from ERC/DRC JSON report files.
-/// Always emits both erc and drc entries (skipped if report is missing/unparseable).
-fn build_manifest_checks(erc_path: &Path, drc_path: &Path) -> Vec<serde_json::Value> {
-    let mut checks = Vec::new();
-
-    // ERC check
-    let erc_check = if erc_path.exists() {
-        if let Ok(content) = fs::read_to_string(erc_path) {
-            match boardflow_kicad::report::ErcReport::parse(&content) {
-                Ok(report) => {
-                    let violations = report.actionable_violations();
-                    let error_count =
-                        violations.iter().filter(|v| v.severity == "error").count() as i32;
-                    let warning_count = violations
-                        .iter()
-                        .filter(|v| v.severity == "warning")
-                        .count() as i32;
-                    let status = if error_count > 0 { "failed" } else { "passed" };
-
-                    let findings: Vec<serde_json::Value> =
-                        report
-                            .sheets
-                            .iter()
-                            .flat_map(|sheet| {
-                                sheet.violations.iter().filter(|v| v.is_actionable()).map(
-                                    move |v| {
-                                        let pos_mm = v.items.first().and_then(|item| {
-                                            item.pos
-                                                .as_ref()
-                                                .map(|p| serde_json::json!({"x": p.x, "y": p.y}))
-                                        });
-                                        let mut finding = serde_json::json!({
-                                            "severity": v.severity,
-                                            "rule_code": v.violation_type,
-                                            "title": v.description,
-                                            "subject_kind": "schematic",
-                                            "sheet_path": sheet.path,
-                                        });
-                                        if let Some(pos) = pos_mm {
-                                            finding
-                                                .as_object_mut()
-                                                .unwrap()
-                                                .insert("pos_mm".to_string(), pos);
-                                        }
-                                        finding
-                                    },
-                                )
-                            })
-                            .collect();
-
-                    serde_json::json!({
-                        "kind": "erc",
-                        "status": status,
-                        "error_count": error_count,
-                        "warning_count": warning_count,
-                        "notice_count": 0,
-                        "tool_name": "kicad-cli",
-                        "findings": findings,
-                    })
-                }
-                Err(_) => serde_json::json!({
-                    "kind": "erc",
-                    "status": "skipped",
-                    "error_count": 0,
-                    "warning_count": 0,
-                    "notice_count": 0,
-                    "findings": [],
-                }),
-            }
-        } else {
-            serde_json::json!({
-                "kind": "erc",
-                "status": "skipped",
-                "error_count": 0,
-                "warning_count": 0,
-                "notice_count": 0,
-                "findings": [],
-            })
-        }
-    } else {
-        serde_json::json!({
-            "kind": "erc",
-            "status": "skipped",
-            "error_count": 0,
-            "warning_count": 0,
-            "notice_count": 0,
-            "findings": [],
-        })
-    };
-    checks.push(erc_check);
-
-    // DRC check
-    let drc_check = if drc_path.exists() {
-        if let Ok(content) = fs::read_to_string(drc_path) {
-            match boardflow_kicad::report::DrcReport::parse(&content) {
-                Ok(report) => {
-                    let violations = report.actionable_violations();
-                    let error_count =
-                        violations.iter().filter(|v| v.severity == "error").count() as i32;
-                    let warning_count = violations
-                        .iter()
-                        .filter(|v| v.severity == "warning")
-                        .count() as i32;
-                    let status = if error_count > 0 { "failed" } else { "passed" };
-
-                    let findings: Vec<serde_json::Value> = report
-                        .all_violations()
-                        .into_iter()
-                        .filter(|v| v.is_actionable())
-                        .map(|v| {
-                            let pos_mm = v.items.first().and_then(|item| {
-                                item.pos
-                                    .as_ref()
-                                    .map(|p| serde_json::json!({"x": p.x, "y": p.y}))
-                            });
-                            let mut finding = serde_json::json!({
-                                "severity": v.severity,
-                                "rule_code": v.violation_type,
-                                "title": v.description,
-                                "subject_kind": "pcb",
-                            });
-                            if let Some(pos) = pos_mm {
-                                finding
-                                    .as_object_mut()
-                                    .unwrap()
-                                    .insert("pos_mm".to_string(), pos);
-                            }
-                            finding
-                        })
-                        .collect();
-
-                    serde_json::json!({
-                        "kind": "drc",
-                        "status": status,
-                        "error_count": error_count,
-                        "warning_count": warning_count,
-                        "notice_count": 0,
-                        "tool_name": "kicad-cli",
-                        "findings": findings,
-                    })
-                }
-                Err(_) => serde_json::json!({
-                    "kind": "drc",
-                    "status": "skipped",
-                    "error_count": 0,
-                    "warning_count": 0,
-                    "notice_count": 0,
-                    "findings": [],
-                }),
-            }
-        } else {
-            serde_json::json!({
-                "kind": "drc",
-                "status": "skipped",
-                "error_count": 0,
-                "warning_count": 0,
-                "notice_count": 0,
-                "findings": [],
-            })
-        }
-    } else {
-        serde_json::json!({
-            "kind": "drc",
-            "status": "skipped",
-            "error_count": 0,
-            "warning_count": 0,
-            "notice_count": 0,
-            "findings": [],
-        })
-    };
-    checks.push(drc_check);
-
-    checks
 }
